+++ gradle/libs.versions.toml
@@ -7,6 +7,7 @@ kotlinxCoroutinesGuavaVersion = "1.8.1"
 leakcanaryAndroidVersion = "2.14"
 lifecycleProcessVersion = "2.8.1"
 mockk = "1.13.8"
+orchestratorVersion = "1.5.0"
 revenue_cat_version = "7.7.1"
 compose_version = "2024.05.00"
 kotlinx_serialization_json_version = "1.6.0"
@@ -21,15 +22,15 @@ lifecycle_runtime_ktx_version = "2.8.1"
 junit_version = "4.13.2"
 kotlinx_coroutines_test_version = "1.8.1"
 room_runtime_version = "2.6.1"
-test_ext_junit_version = "1.1.5"
+test_ext_junit_version = "1.2.0"
 espresso_core_version = "3.5.1"
-test_runner_version = "1.4.0"
-test_rules_version = "1.4.0"
+test_runner_version = "1.6.0"
+test_rules_version = "1.6.0"
 kotlin = "1.9.0"
 kotlinx_coroutines_core_version = "1.8.1"
 mockk_version = "1.12.8"
 threetenbp_version = "1.6.8"
-uiautomator_version = "2.3.0-alpha03"
+uiautomator_version = "2.3.0"
 workRuntimeKtx_version = "2.9.0"
 serialization_version = "1.6.0"
 dropshot_version = "0.4.2"
@@ -40,6 +41,7 @@ dropshot_version = "0.4.2"
 
 javascriptengine = { module = "androidx.javascriptengine:javascriptengine", version.ref = "javascriptengineVersion" }
 leakcanary-android = { module = "com.squareup.leakcanary:leakcanary-android", version.ref = "leakcanaryAndroidVersion" }
+orchestrator = { module = "androidx.test:orchestrator", version.ref = "orchestratorVersion" }
 room-compiler = { module = "androidx.room:room-compiler", version.ref = "room_runtime_version" }
 room-ktx = { module = "androidx.room:room-ktx", version.ref = "room_runtime_version" }
 room-runtime = { module = "androidx.room:room-runtime", version.ref = "room_runtime_version" }
